motion-password-server/requirements.txt
fastapi
websockets
uvicorn
httpx
websocket
websocket-client
sqlalchemy
python-multipart
pandas
lightning
torchvision
pytest
pytorch-metric-learning
psutil
matplotlib
seaborn
tabulate
motion-learning-toolbox
pytorch-lightning